Revised returned product and fix returned product and products.

+++ Forms/frmListProducts.cs
@@ -20,6 +20,53 @@ namespace RAloverasPharmacyPOSSystem.Forms
         Components.Value val = new Components.Value();
         Functions.Product product = new Functions.Product();
 
+        bool isProducts = false;
+        bool isReturnedProducts = false;
+
+        private void LoadProductsAction()
+        {
+            if (val.MyUserLevel == "ADMIN")
+            {
+                DataGridViewButtonColumn btnUpdate = new DataGridViewButtonColumn();
+                btnUpdate.HeaderText = "ACTION";
+                btnUpdate.Name = "btnUpdate";
+                btnUpdate.Text = "UPDATE";
+                btnUpdate.UseColumnTextForButtonValue = true;
+                this.gridProducts.Columns.Insert(0, btnUpdate);
+
+                DataGridViewButtonColumn btnDelete = new DataGridViewButtonColumn();
+                btnDelete.HeaderText = "";
+                btnDelete.Name = "btnDelete";
+                btnDelete.Text = "DELETE";
+                btnDelete.UseColumnTextForButtonValue = true;
+                this.gridProducts.Columns.Insert(1, btnDelete);
+            }
+        }
+
+        private void LoadProducts()
+        {
+            isReturnedProducts = false;
+            isProducts = true;
+
+            this.gridProducts.Columns.Clear();
+            product.LoadProducts(this.gridProducts);
+            LoadProductsAction();
+
+            this.gridProducts.ClearSelection();
+        }
+
+        private void LoadReturnedProducts()
+        {
+            isProducts = false;
+            isReturnedProducts = true;
+
+            this.gridProducts.Columns.Clear();
+            product.LoadReturnedProducts(this.gridProducts);
+            LoadProductsAction();
+
+            this.gridProducts.ClearSelection();
+        }
+
         private void OpenAddProductForm()
         {
             this.gridProducts.ClearSelection();
@@ -28,7 +75,8 @@ namespace RAloverasPharmacyPOSSystem.Forms
             addProduct.ShowDialog();
         }
 
-        private void OpenAddReturnProductForm() {
+        private void OpenAddReturnProductForm()
+        {
             Forms.frmDashboard dashboard = (Forms.frmDashboard)Application.OpenForms["frmDashboard"];
             Panel pnlMain = (Panel)dashboard.Controls["pnlMain"];
 
@@ -47,29 +95,75 @@ namespace RAloverasPharmacyPOSSystem.Forms
 
         private void frmListProducts_KeyDown(object sender, KeyEventArgs e)
         {
-            if (e.KeyCode == Keys.F1) {
+            if (e.KeyCode == Keys.F1)
+            {
                 OpenAddProductForm();
-            } else if(e.KeyCode == Keys.F2) {
+            } 
+            else if(e.KeyCode == Keys.F2)
+            {
                 OpenAddReturnProductForm();
             }
         }
 
         private void gridProducts_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
-            if(this.gridProducts.Columns[e.ColumnIndex].Name == "btnUpdate") {
-                if(product.GetProduct(long.Parse(this.gridProducts.SelectedCells[2].Value.ToString()))) {
+            if(isProducts && this.gridProducts.Columns[e.ColumnIndex].Name == "btnUpdate") 
+            {
+                if(product.GetProduct(long.Parse(this.gridProducts.SelectedCells[2].Value.ToString())))
+                {
                     Forms.frmUpdateProduct updateProduct = new Forms.frmUpdateProduct();
                     updateProduct.ShowDialog();
                 }
             }
-            else if(this.gridProducts.Columns[e.ColumnIndex].Name == "btnDelete") {
+            else if(isProducts && this.gridProducts.Columns[e.ColumnIndex].Name == "btnDelete") 
+            {
                 if(MessageBox.Show("Are you sure you want to delete this product?", "", MessageBoxButtons.YesNo, MessageBoxIcon.Question)
-                    == DialogResult.Yes) {
-                    if(product.DeleteProduct(long.Parse(this.gridProducts.SelectedCells[2].Value.ToString()))) {
+                    == DialogResult.Yes)
+                {
+                    if(product.DeleteProduct(long.Parse(this.gridProducts.SelectedCells[2].Value.ToString()))) 
+                    {
                         MessageBox.Show("Product was successfully deleted!", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
                         product.LoadProducts(this.gridProducts);
-                    } else {
+                    } 
+                    else 
+                    {
                         MessageBox.Show("Failed to delete prodct!", "", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                }
+
+                this.gridProducts.ClearSelection();
+                this.txtSearch.Focus();
+            }
+            else if(isReturnedProducts && this.gridProducts.Columns[e.ColumnIndex].Name == "btnUpdate")
+            {
+                if(product.GetReturnedProduct(long.Parse(this.gridProducts.SelectedCells[2].Value.ToString())))
+                {
+                    Forms.frmDashboard dashboard = (Forms.frmDashboard)Application.OpenForms["frmDashboard"];
+                    Panel pnlMain = (Panel)dashboard.Controls["pnlMain"];
+                    
+                    pnlMain.Controls.Clear();
+                    Forms.frmUpdateReturnProduct updateReturnProduct = new Forms.frmUpdateReturnProduct();
+                    updateReturnProduct.TopLevel = false;
+                    pnlMain.Controls.Add(updateReturnProduct);
+                    updateReturnProduct.Dock = DockStyle.Fill;
+                    updateReturnProduct.Show();
+                }
+            }
+            else if(isReturnedProducts && this.gridProducts.Columns[e.ColumnIndex].Name == "btnDelete")
+            {
+                if (MessageBox.Show("Are you sure you want to delete this returned product?", "", MessageBoxButtons.YesNo, MessageBoxIcon.Question)
+                    == DialogResult.Yes) 
+                { 
+                    if(product.DeleteReturnedProduct(long.Parse(this.gridProducts.SelectedCells[2].Value.ToString()))) 
+                    {
+                        MessageBox.Show("Returned product was successfully deleted!", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    
+                        product.LoadReturnedProducts(this.gridProducts);
+                        this.gridProducts.ClearSelection();
+                    } 
+                    else 
+                    {
+                        MessageBox.Show("Failed to delete returned product!", "", MessageBoxButtons.OK, MessageBoxIcon.Error);
                     }
                 }
 
@@ -80,9 +174,12 @@ namespace RAloverasPharmacyPOSSystem.Forms
 
         private void txtSearch_TextChanged(object sender, EventArgs e)
         {
-            if(String.IsNullOrWhiteSpace(this.txtSearch.Text)) {
+            if(String.IsNullOrWhiteSpace(this.txtSearch.Text)) 
+            {
                 product.LoadProducts(this.gridProducts);
-            } else {
+            }
+            else
+            {
                 product.SearchProduct(this.txtSearch.Text, this.gridProducts);
             }
 
@@ -99,23 +196,7 @@ namespace RAloverasPharmacyPOSSystem.Forms
             this.KeyPreview = true;
             this.txtSearch.CharacterCasing = CharacterCasing.Upper;
 
-            product.LoadProducts(this.gridProducts);
-
-            if(val.MyUserLevel == "ADMIN") {
-                DataGridViewButtonColumn btnUpdate = new DataGridViewButtonColumn();
-                btnUpdate.HeaderText = "ACTION";
-                btnUpdate.Name = "btnUpdate";
-                btnUpdate.Text = "UPDATE";
-                btnUpdate.UseColumnTextForButtonValue = true;
-                this.gridProducts.Columns.Insert(0, btnUpdate);
-
-                DataGridViewButtonColumn btnDelete = new DataGridViewButtonColumn();
-                btnDelete.HeaderText = "";
-                btnDelete.Name = "btnDelete";
-                btnDelete.Text = "DELETE";
-                btnDelete.UseColumnTextForButtonValue = true;
-                this.gridProducts.Columns.Insert(1, btnDelete);
-            }
+            LoadProducts();
 
             this.txtSearch.Focus();
         }
@@ -128,6 +209,16 @@ namespace RAloverasPharmacyPOSSystem.Forms
         private void btnAddReturnProduct_Click(object sender, EventArgs e)
         {
             OpenAddReturnProductForm();
+        }
+
+        private void btnLoadProducts_Click(object sender, EventArgs e)
+        {
+            LoadProducts();
+        }
+
+        private void btnLoadReturnedProducts_Click(object sender, EventArgs e)
+        {
+            LoadReturnedProducts();
         }
 
         private void btnNext_Click(object sender, EventArgs e)
